(#5653) - automatically determine dependencies at publish time
