+++ package.json
@@ -47,7 +47,6 @@
     "localstorage-down": "0.6.6",
     "ltgt": "2.1.2",
     "memdown": "1.2.1",
-    "pull-stream": "3.4.5",
     "readable-stream": "1.0.33",
     "request": "2.74.0",
     "scope-eval": "0.0.3",
@@ -62,6 +61,7 @@
     "body-parser": "1.15.2",
     "browserify": "13.1.0",
     "browserify-incremental": "3.1.1",
+    "builtin-modules": "1.1.1",
     "bundle-collapser": "1.2.1",
     "chai": "3.5.0",
     "chai-as-promised": "5.3.0",
@@ -74,6 +74,8 @@
     "eslint": "3.4.0",
     "express": "4.14.0",
     "express-pouchdb": "1.0.7",
+    "find-requires": "0.2.2",
+    "glob": "7.0.6",
     "http-server": "0.9.0",
     "istanbul": "0.4.5",
     "istanbul-coveralls": "1.0.3",
@@ -81,6 +83,8 @@
     "less": "2.7.1",
     "lie": "3.1.0",
     "lodash.debounce": "4.0.8",
+    "lodash.flatten": "4.4.0",
+    "lodash.uniq": "4.5.0",
     "mkdirp": "0.5.1",
     "mocha": "2.5.3",
     "mockery": "1.7.0",
